# Perf: remove Onyx usage from `withLocalize`