<!-- If necessary, assign reviewers that know the area or changes well. Feel free to tag any additional reviewers you see fit. -->
@marcaaron 

### Details
<!-- Explanation of the change or anything fishy that is going on -->
This component is one of the most used components and has more than 310 instances depending on chat size
Instead of creating a new connection for each `withLocalize` usage - use the already cached value in `translate`
This significantly reduces the `Onyx.get` load that is caused by all the connections

### Fixed Issues (Related To)
Fixes #4268

### Tests
<!--- 
Add a numbered list of manual tests you performed that validates your changes work on all platforms, and that there are no regressions present.
Add any additional test steps if test steps are unique to a particular platform.
Manual test steps should be written so that your reviewer can repeat and verify one or more expected outcomes in the development environment.

For example:
1. Click on the text input to bring it into focus
2. Upload an image via copy paste
3. Verify a modal appears displaying a preview of that image 
--->
1. Measure performance (init time) before this change
2. Measure performance (init time) with the change in this PR

I'm using `Onyx.printMetrics()` to monitor the `Onyx.get` calls that happen during init: https://github.com/Expensify/react-native-onyx#benchmarks

#### My Benchmarks

##### Before
Total: | 123.80sec |   |   | Last | 17.49sec |   |   | ReportID | 71955477
-- | -- | -- | -- | -- | -- | -- | -- | -- | --

method | total time spent | max | min | avg | time last call completed | calls made
-- | -- | -- | -- | -- | -- | --
Onyx:get | 96.85sec | 2.92sec | 0.10ms | 144.77ms | 17.48sec | 669
- see the individual Onyx.get calls here: https://docs.google.com/spreadsheets/d/1wGrVXsad6ayQZVqvyUmTuTb1cmeegdE0y-IFzhLiPKU/edit?usp=sharing
  - search for `preferredLocale` 

##### After
Total: | 62.62sec |   |   | Last | 12.27sec |   |   | ReportID | 71955477
-- | -- | -- | -- | -- | -- | -- | -- | -- | --

method | total time spent | max | min | avg | time last call completed | calls made
-- | -- | -- | -- | -- | -- | --
Onyx:get | 40.44sec | 1.79sec | 0.08ms | 139.94ms | 12.27sec | 289
- excel data: https://docs.google.com/spreadsheets/d/1cVe3eEnwYgwbH4DRV__g9_W8kQFpVhvTjDu3Igr6Ds0/edit?usp=sharing

### QA Steps
<!--- 
Add a numbered list of manual tests that can be performed by our QA engineers on the staging environment to validate that your changes work on all platforms, and that there are no regressions present.
Add any additional QA steps if test steps are unique to a particular platform.
Manual test steps should be written so that the QA engineer can repeat and verify one or more expected outcomes in the staging environment.

For example:
1. Click on the text input to bring it into focus
2. Upload an image via copy paste
3. Verify a modal appears displaying a preview of that image 
--->

### Tested On

- [x] Web
- [ ] Mobile Web
- [ ] Desktop
- [x] iOS
- [x] Android

### Screenshots
<!-- Add screenshots for all platforms tested. Pull requests won't be merged unless the screenshots show the app was tested on all platforms.-->

#### Web
<!-- Insert screenshots of your changes on the web platform-->
![image](https://user-images.githubusercontent.com/12156624/127711858-844f638d-6950-4402-ad4f-49770a807923.png)

#### Mobile Web
<!-- Insert screenshots of your changes on the web platform (from a mobile browser)-->

#### Desktop
<!-- Insert screenshots of your changes on the desktop platform-->

#### iOS
<!-- Insert screenshots of your changes on the iOS platform-->

#### Android
<!-- Insert screenshots of your changes on the Android platform-->