@Julesssss 2021-07-27
> When we update the locale, Will the app re render components. I don't think it will.

That's a good point. I wouldn't have thought so either.