@kidroca 2021-07-27
> When we update the locale, Will the app re render components. I don't think it will.

When does that happen, isn't the app reloading when a new locale is selected?