@iwiznia 2021-07-27
> When does that happen, isn't the app reloading when a new locale is selected?

You can test this by changing the language in the preferences (maybe you need to force being in the bet if that's not done by default in dev). And no, it does not reload, all components that are using withLocalize will re-render themselves.