@kidroca 2021-07-27
I've mentioned this at Slack, there's another way to implement the HOC and it will not suffer this problem: https://expensify.slack.com/archives/C01GTK53T8Q/p1627403864349200?thread_ts=1627393814.334100&cid=C01GTK53T8Q

The changes would be slightly more, but it might be an even faster if the HOC is implemented through `React.createContext`. As this will also save 300+ class instances of `withLocalize`

I'll try it out tomorrow and do another benchmark