@marcaaron 2021-07-27
Looks good so far and definitely up for more improvements. I'm adding a link back to the original issue in the description so there is some more context.