@kidroca 2021-07-27
> But I think maybe this needs a new issue?

Yeah, I'll probably have to spent some more time here, and then test/bench the changes on the remaining platforms.

I can open an issue tomorrow