@kidroca 2021-07-30
Updated implementation and posted new benchmarks here: https://github.com/Expensify/App/issues/4268#issuecomment-890072411

I think another possible optimization that might be applied at the moment is to `drop` the `withOnyx` call altogether like this:

```jsx
    componentDidMount() {
        this.connectionId = Onyx.connect({
            key: ONYXKEYS.NVP_PREFERRED_LOCALE,
            callback: (val) => {
                if (val) {
                    this.setState({preferredLocale: val});
                }
            },
        });
    }

    componentWillUnmount() {
        Onyx.disconnect(this.connectionId);
    }
```

This skips creating the `withOnyx` class component
I've tried it briefly it was slightly faster bringing "Last call ended time" to ~10.87s (otherwise it's ~11.05s) 
I can try to capture more data about it next week, but it might be better to focus/find something more important