@marcaaron 2021-07-30
>This skips creating the withOnyx class component
>I've tried it briefly it was slightly faster bringing "Last call ended time" to ~10.87s (otherwise it's ~11.05s)
>I can try to capture more data about it next week, but it might be better to focus/find something more important

Yeah, I think there are probably better optimizations to focus on. This is already so great 🎉