@kidroca 2021-07-31
@isagoico No specific test steps other than basic regressions as translations are used everywhere. 
Switching the app language should work