@nkuoch 2021-08-02
@kidroca this PR created [this regression](https://github.com/Expensify/App/issues/4352#issuecomment-890863100).
Seems like [it was pointed out](https://github.com/Expensify/App/pull/4253#issuecomment-887555934) before the merge though. Not sure why we merged it regardless and what was decided about it though? cc @marcaaron