@iwiznia 2021-08-02
> Are we going to handle this as a continuation PR that fixes the regression?

IMO yes, because it's a feature that worked and was broken by this PR.