@kidroca 2021-08-02
Shouldn't I be assigned to the new ticket as well then?