@iwiznia 2021-08-02
It makes it easier to find things if they are assigned to the people that worked on them. It is also used in some of our internal tooling to for example assign reviewers for issues or PRs.