@marcaaron 2021-08-02
>Not sure why we merged it regardless and what was decided about it though? 
 
Oh, hmm, we merged it because the implementation was entirely different. In short, we didn't suspect the new one would have the same issue.